requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
requests>=2.28.0
newspaper3k>=0.2.8
sumy>=0.9.0
feedparser>=6.0.11
beautifulsoup4>=4.12.0
